--- src/components/Home/home_component.js
@@ -1,1 +1,0 @@
-vari="home";
